refactor(Team): prefer dedicated data folder
