+++ src/routes/people/+page.server.ts
@@ -1,5 +1,5 @@
 import { getExec } from '$lib/people/exec/exec';
-import { getTeam } from '$lib/people/team/team';
+import { getTeam } from '$lib/data/team';
 
 export async function load() {
     return { team: await getTeam(), exec: (await getExec()).reverse() };
